backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
boto3==1.26.86
botocore==1.29.86
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
ipython==8.11.0
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
multidict==6.0.4
mysqlclient==2.1.1
openai==0.27.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
yarl==1.8.2
